feat(restapi): moving validation service to entrypoints
