+++ src/dioptra/restapi/v1/entrypoints/service.py
@@ -27,8 +27,12 @@ from structlog.stdlib import BoundLogger
 
 from dioptra.restapi.db import db, models
 from dioptra.restapi.db.models.constants import resource_lock_types
+from dioptra.restapi.db.models.plugins import (
+    PluginTaskOutputParameter,
+)
 from dioptra.restapi.errors import (
     BackendDatabaseError,
+    EmptyGraphError,
     EntityDoesNotExistError,
     EntityExistsError,
     EntrypointSwapsRenderError,
@@ -49,8 +53,12 @@ from dioptra.restapi.v1.shared.task_engine_yaml.service import (
     TaskEngineYamlService,
     coerce_entrypoint_default_param_types,
 )
+from dioptra.restapi.v1.shared.views import (
+    get_plugin_plugin_files_from_plugin_snapshot_ids,
+)
 from dioptra.sdk.utilities.entrypoint_swaps import render_swaps_graph
 from dioptra.task_engine import util
+from dioptra.task_engine.issues import IssueSeverity, IssueType, ValidationIssue
 
 LOGGER: BoundLogger = structlog.stdlib.get_logger()
 PLUGIN_RESOURCE_TYPE: Final[str] = "entry_point_plugin"
@@ -81,6 +89,7 @@ class EntrypointService(object):
         plugin_ids_service: PluginIdsService,
         queue_ids_service: QueueIdsService,
         group_id_service: GroupIdService,
+        swaps_validation_service: "SwapsValidationService",
     ) -> None:
         """Initialize the entrypoint service.
 
@@ -96,6 +105,7 @@ class EntrypointService(object):
         self._plugin_ids_service = plugin_ids_service
         self._queue_ids_service = queue_ids_service
         self._group_id_service = group_id_service
+        self._swaps_validation_service = swaps_validation_service
 
     def create(
         self,
@@ -141,8 +151,6 @@ class EntrypointService(object):
         """
         log: BoundLogger = kwargs.get("log", LOGGER.new())
 
-        # TODO: need to add a check here that graphs are valid yaml
-
         duplicate = self._entrypoint_name_service.get(name, group_id=group_id, log=log)
         if duplicate is not None:
             raise EntityExistsError(
@@ -173,6 +181,17 @@ class EntrypointService(object):
             creator=current_user,
         )
         db.session.add(new_entrypoint)
+
+        # Validate swaps graph before committing the entrypoint
+        self._swaps_validation_service.validate(
+            group_id=group_id,
+            swaps_graph=task_graph,
+            artifact_graph=artifact_graph,
+            entrypoint_parameters=parameters,
+            entrypoint_artifacts=artifact_parameters,
+            plugin_snapshot_ids=plugin_ids,
+            log=log,
+        )
 
         new_entrypoint.entry_point_plugins = [
             models.EntryPointPlugin(
@@ -1688,6 +1707,355 @@ class DynamicGlobalParametersService(object):
                         )  # if it is not a step output, it must be a global param
 
         return needed_vars, used_tasks
+
+
+class SwapsValidationService(object):
+    """Initialize the swaps validation resource.
+
+    All arguments are provided via dependency injection.
+
+    Args:
+        entrypoint_snapshot_id_service: An EntrypointSnapshotIdService object.
+        task_engine_yaml_service: A TaskEngineYamlService object.
+        dynamic_global_parameter_service: A DynamicGlobalParametersService object.
+    """
+
+    @inject
+    def __init__(
+        self,
+        entrypoint_snapshot_id_service: EntrypointSnapshotIdService,
+        task_engine_yaml_service: TaskEngineYamlService,
+        dynamic_global_parameter_service: DynamicGlobalParametersService,
+    ) -> None:
+        self._entrypoint_snapshot_id_service = entrypoint_snapshot_id_service
+        self._task_engine_yaml_service = task_engine_yaml_service
+        self._dynamic_global_parameter_service = dynamic_global_parameter_service
+
+    def swaps_graph_validation(
+        self,
+        pre_rendered_task_graph: dict[str, Any],
+    ) -> list[ValidationIssue]:
+        from dioptra.sdk.api.swappable_validation import get_json_schema
+        from dioptra.task_engine.validation import _schema_validate
+
+        return _schema_validate(pre_rendered_task_graph, get_json_schema())
+
+    def validate_swap_output_matches(
+        self, pre_rendered_task_graph: dict[str, Any], task_lookup_dict: dict[str, Any]
+    ) -> tuple[list[ValidationIssue], dict[str, Any]]:
+        """Assess whether the output types of all tasks within a swap match.
+
+        Args:
+            pre_rendered_task_graph: The task graph dictionary, before it is rendered.
+            task_lookup_dict: A dictionary mapping tasks to their plugins, generated by build_task_lookup_dict.
+
+        Returns:
+            A list of validation issues and tasks needed for the graph.
+        """
+
+        mismatched_aliases = {}
+        swap_tasks: dict[str, Any] = {}
+        collected_no_tasks_found = []
+
+        for _step, task in pre_rendered_task_graph.items():
+            if isinstance(task, dict):
+                for swap_name, aliased_defns in task.items():
+                    if swap_name.startswith("?"):
+                        output_types = set()
+                        swap_dict = {}
+
+                        for alias, definition in aliased_defns.items():
+                            # figure out whether its long or short form
+                            if "task" in definition:
+                                # long version
+                                task_name = definition["task"]
+                            else:
+                                # short version - should be exactly one key in here
+                                task_name = list(definition.keys())[0]
+
+                            # get the output type of task_name in the plugin
+                            if task_name in task_lookup_dict:
+                                swap_dict[alias] = {
+                                    "plugin_snapshot_id": task_lookup_dict[task_name][
+                                        "plugin_snapshot_id"
+                                    ],
+                                    "pluginfile_filename": task_lookup_dict[task_name][
+                                        "pluginfile_filename"
+                                    ],
+                                    "task_name": task_lookup_dict[task_name][
+                                        "task_name"
+                                    ],
+                                }
+
+                                output_parameters: list[PluginTaskOutputParameter] = (
+                                    task_lookup_dict[task_name]["output_parameters"]
+                                )
+                                output_types.add(
+                                    tuple(
+                                        [
+                                            parameter.parameter_type.name
+                                            for parameter in output_parameters
+                                        ]
+                                    )
+                                )
+                            else:
+                                collected_no_tasks_found.append(
+                                    ValidationIssue(
+                                        type_=IssueType.SEMANTIC,
+                                        severity=IssueSeverity.ERROR,
+                                        message=f"In swap '{swap_name}', task with name '{task_name}' not found in registered tasks.",
+                                    )
+                                )
+
+                        swap_tasks[swap_name] = swap_dict
+
+                        if len(output_types) > 1:
+                            mismatched_aliases[swap_name] = output_types
+
+        return collected_no_tasks_found + [
+            ValidationIssue(
+                type_=IssueType.TYPE,
+                severity=IssueSeverity.ERROR,
+                message=f"Swap '{swap_name}' contains mismatched output types: {types}",
+            )
+            for swap_name, types in mismatched_aliases.items()
+        ], swap_tasks
+
+    def build_task_lookup_dict(
+        self, plugins: list[models.PluginPluginFile]
+    ) -> dict[str, Any]:
+        """Generates a lookup dictionary for function tasks.
+
+        Args:
+            plugins: A list of PluginPluginFile objects.
+
+        Returns:
+            A dictionary mapping function tasks to their plugin snapshot IDs, plugin filenames,
+            task names, and output parameters.
+        """
+        lookup = {}
+
+        for pair in plugins:
+            plugin = pair.plugin
+            file = pair.plugin_file
+
+            for task in file.tasks:
+                if isinstance(task, models.FunctionTask):
+                    # there must be a better way to check this
+
+                    lookup[task.plugin_task_name] = {
+                        "plugin_snapshot_id": plugin.resource_snapshot_id,
+                        "pluginfile_filename": file.filename,
+                        "task_name": task.plugin_task_name,
+                        "output_parameters": task.output_parameters,
+                    }
+
+        return lookup
+
+    def extract_swaps(self, task_graph: dict[str, Any]) -> dict[str, list[str]]:
+        """Extract all swaps from a task graph.
+
+        Args:
+            task_graph: The task graph dictionary.
+
+        Returns:
+            A dictionary mapping swap names to lists of available aliases.
+        """
+        swaps = {}
+
+        for _step, task in task_graph.items():
+            for swap_name, aliased_defns in task.items():
+                if swap_name.startswith("?"):
+                    swap_name_clean = swap_name[1:]  # Remove the '?' prefix
+                    if swap_name_clean not in swaps:
+                        swaps[swap_name_clean] = list(aliased_defns.keys())
+
+        return swaps
+
+    def validate_single_swap_combinations(
+        self,
+        task_graph_yaml: dict[str, Any],
+        artifact_graph: str,
+        plugin_plugin_files: list[models.PluginPluginFile],
+        plugin_parameter_types: list[models.PluginTaskParameterType],
+        entrypoint_parameters: list[dict[str, Any]],
+        entrypoint_artifacts: list[dict[str, Any]],
+        swap_choices: dict[str, Any],
+        log: BoundLogger,
+    ) -> tuple[list[ValidationIssue], set[str]]:
+        """Renders a graph containing swaps for a given set of swap_choices, and performs
+        validation on the rendered graph.
+
+        Args:
+            task_graph_yaml: The task graph dictionary.
+            artifact_graph: The artifact graph dictionary.
+            plugin_plugin_files: A list of Plugin-PluginFile objects.
+            plugin_parameter_types: A list of plugin parameter types.
+            entrypoint_parameters: A list of entrypoint parameters.
+            entrypoint_artifacts: A list of entrypoint artifacts parameters.
+            swap_choices: A dictionary representing the choices for each swap.
+            log: A BoundLogger instance.
+
+        Returns:
+            A list of validation issues and a list of global parameters required for the graph.
+        """
+        import json
+
+        from dioptra.restapi.v1.shared.entrypoint_validation import (
+            build_entrypoint_data_adapter,
+        )
+
+        issues_for_swap = []
+        required_globals: set[str] = set()
+
+        try:
+            rendered_graph = render_swaps_graph(task_graph_yaml, swap_choices)
+
+            entrypoint_data = build_entrypoint_data_adapter(
+                json.dumps(rendered_graph),
+                artifact_graph,
+                entrypoint_parameters,
+                entrypoint_artifacts,
+                log,
+            )
+            task_engine_dict = self._task_engine_yaml_service.build_dict(
+                entry_point=entrypoint_data,
+                plugin_plugin_files=plugin_plugin_files,
+                plugin_parameter_types=plugin_parameter_types,
+            )
+            # this is a schema check and deeper validation no longer present in workflows
+            issues = self._task_engine_yaml_service.validate(task_engine_dict)
+
+            for issue in issues:
+                issue.message = f"[Swap combination {swap_choices}] {issue.message}"
+                issues_for_swap.append(issue)
+
+            # collect any globals needed for this rendering
+            required_globals, _ = (
+                self._dynamic_global_parameter_service.get_required_globals(
+                    rendered_graph
+                )
+            )
+
+        except Exception as e:
+            issues_for_swap.append(
+                ValidationIssue(
+                    type_=IssueType.SEMANTIC,
+                    severity=IssueSeverity.ERROR,
+                    message=f"[Swap combination {swap_choices}] Error rendering graph: {str(e)}",
+                )
+            )
+
+        return issues_for_swap, required_globals
+
+    def validate(
+        self,
+        group_id: int,
+        swaps_graph: str,
+        artifact_graph: str,
+        entrypoint_parameters: list[dict[str, Any]],
+        entrypoint_artifacts: list[dict[str, Any]],
+        plugin_snapshot_ids: list[int],
+        **kwargs,
+    ) -> dict[str, Any]:
+        """Validation for a proposed entrypoint which may contain swaps, looping over swap combinations.
+
+        Args:
+            group_id: The group ID.
+            swaps_graph: The task graph dictionary.
+            artifact_graph: The artifact graph dictionary.
+            entrypoint_parameters: A list of entrypoint parameters.
+            entrypoint_artifacts: A list of entrypoint artifacts parameters.
+            plugin_snapshot_ids: A list of plugin snapshot IDs needed for the entrypoint.
+
+        Returns:
+            A ValidateSwapsResponseSchema object.
+        """
+
+        log: BoundLogger = kwargs.get("log", LOGGER.new())
+
+        print(swaps_graph, flush=True)
+        swaps_yaml = yaml.safe_load(swaps_graph)
+
+        if swaps_yaml is None:
+            raise EmptyGraphError("Provided swaps graph is empty.")
+
+        # check for pre-render schema issues
+        schema_issues = self.swaps_graph_validation(pre_rendered_task_graph=swaps_yaml)
+        schema_valid = len(schema_issues) == 0
+
+        output_issues = []
+        collected_rendered_validation_issues = []
+        collected_required_globals = set()
+
+        if schema_valid:
+            # perform swap output matching
+            plugin_plugin_files = get_plugin_plugin_files_from_plugin_snapshot_ids(
+                plugin_snapshot_ids=plugin_snapshot_ids, logger=log
+            )
+
+            # build a lookup dictionary for tasks from the plugin files
+            task_lookup_dict = self.build_task_lookup_dict(plugin_plugin_files)
+
+            # validate that swap outputs match across the graph
+            output_issues, _ = self.validate_swap_output_matches(
+                pre_rendered_task_graph=swaps_yaml,
+                task_lookup_dict=task_lookup_dict,
+            )
+            # extract a mapping of swaps to possible swap choices
+            swaps = self.extract_swaps(
+                swaps_yaml
+            )  # { swap1: [alias1, alias2, alias3], swap2: [alias4, alias5, alias6], etc. }
+            swap_names = list(swaps.keys())  # [swap1, swap2, etc.]
+
+            combinations = []
+
+            # create a list of swap choice combinations to try
+            for swap_name in swap_names:
+                for alias in swaps[swap_name]:
+                    current_swap_choice = {
+                        s: swaps[s][0] for s in swaps.keys() if s != swap_name
+                    }  # keeping the first swap from every *other* swap, as we vary the *current* swap, O(n) instead of O(w*x*y*z*...)!
+                    current_swap_choice[swap_name] = alias
+
+                    combinations.append(current_swap_choice)
+
+            combinations = [{}] if combinations == [] else combinations  #
+
+            # loop over the combinations, render each, and validate as a normal experiment description
+            for combination in combinations:
+                rendered_validation_issues, required_globals = (
+                    self.validate_single_swap_combinations(
+                        artifact_graph=artifact_graph,
+                        task_graph_yaml=swaps_yaml,
+                        plugin_plugin_files=plugin_plugin_files,
+                        plugin_parameter_types=self._entrypoint_snapshot_id_service.get_group_plugin_parameter_types(
+                            group_id=group_id, logger=log
+                        ),
+                        entrypoint_parameters=entrypoint_parameters,
+                        entrypoint_artifacts=entrypoint_artifacts,
+                        swap_choices=combination,
+                        log=log,
+                    )
+                )
+                collected_required_globals.update(required_globals)
+                collected_rendered_validation_issues.extend(rendered_validation_issues)
+
+        # compile a set of all declared parameter names
+        declared_globals = {p["name"] for p in entrypoint_parameters}
+        declared_globals.update({a["name"] for a in entrypoint_artifacts})
+
+        # any variable we have used that is not declared is missing
+        missing_globals = [
+            g for g in collected_required_globals if g not in declared_globals
+        ]
+
+        return {
+            "schema_issues": schema_issues,
+            "swap_issues": output_issues,
+            "rendered_validation_errors": collected_rendered_validation_issues,
+            "missing_global_params": missing_globals,
+        }
 
 
 def _get_entrypoint_plugin_snapshots(
