+++ src/dioptra/restapi/v1/shared/entrypoint_swaps/service.py
@@ -1,420 +1,1 @@
-# This Software (Dioptra) is being made available as a public service by the
-# National Institute of Standards and Technology (NIST), an Agency of the United
-# States Department of Commerce. This software was developed in part by employees of
-# NIST and in part by NIST contractors. Copyright in portions of this software that
-# were developed by NIST contractors has been licensed or assigned to NIST. Pursuant
-# to Title 17 United States Code Section 105, works of NIST employees are not
-# subject to copyright protection in the United States. However, NIST may hold
-# international copyright in software created by its employees and domestic
-# copyright (or licensing rights) in portions of software that were assigned or
-# licensed to NIST. To the extent that NIST holds copyright in this software, it is
-# being made available under the Creative Commons Attribution 4.0 International
-# license (CC BY 4.0). The disclaimers of the CC BY 4.0 International
-# license (CC BY 4.0). The disclaimers of the CC BY 4.0 license apply to all parts
-# of the software developed or licensed by NIST.
-#
-# ACCESS THE FULL CC BY 4.0 LICENSE HERE:
-# https://creativecommons.org/licenses/by/4.0/legalcode
-from typing import Any
-
-import structlog
-import yaml
-from injector import inject
-from structlog.stdlib import BoundLogger
-
-from dioptra.restapi.db import models
-from dioptra.restapi.db.models.plugins import (
-    PluginTaskOutputParameter,
-)
-from dioptra.restapi.errors import EmptyGraphError
-from dioptra.restapi.v1.entrypoints.service import (
-    DynamicGlobalParametersService,
-    EntrypointIdService,
-    EntrypointSnapshotIdService,
-)
 from dioptra.restapi.v1.shared.task_engine_yaml.service import TaskEngineYamlService
-from dioptra.restapi.v1.shared.views import (
-    get_plugin_plugin_files_from_plugin_snapshot_ids,
-)
-from dioptra.sdk.utilities.entrypoint_swaps import render_swaps_graph
-from dioptra.task_engine.issues import IssueSeverity, IssueType, ValidationIssue
-
-LOGGER: BoundLogger = structlog.stdlib.get_logger()
-
-
-class SwapsValidationService(object):
-    """Initialize the swaps validation resource.
-
-    All arguments are provided via dependency injection.
-
-    Args:
-        entrypoint_snapshot_id_service: An EntrypointSnapshotIdService object.
-        task_engine_yaml_service: A TaskEngineYamlService object.
-        dynamic_global_parameter_service: A DynamicGlobalParametersService object.
-    """
-
-    @inject
-    def __init__(
-        self,
-        entrypoint_snapshot_id_service: EntrypointSnapshotIdService,
-        task_engine_yaml_service: TaskEngineYamlService,
-        dynamic_global_parameter_service: DynamicGlobalParametersService,
-    ) -> None:
-        self._entrypoint_snapshot_id_service = entrypoint_snapshot_id_service
-        self._task_engine_yaml_service = task_engine_yaml_service
-        self._dynamic_global_parameter_service = dynamic_global_parameter_service
-
-    def swaps_graph_validation(
-        self,
-        pre_rendered_task_graph: dict[str, Any],
-    ) -> list[ValidationIssue]:
-        from dioptra.sdk.api.swappable_validation import get_json_schema
-        from dioptra.task_engine.validation import _schema_validate
-
-        return _schema_validate(pre_rendered_task_graph, get_json_schema())
-
-    def validate_swap_output_matches(
-        self, pre_rendered_task_graph: dict[str, Any], task_lookup_dict: dict[str, Any]
-    ) -> tuple[list[ValidationIssue], dict[str, Any]]:
-        """Assess whether the output types of all tasks within a swap match.
-
-        Args:
-            pre_rendered_task_graph: The task graph dictionary, before it is rendered.
-            task_lookup_dict: A dictionary mapping tasks to their plugins, generated by build_task_lookup_dict.
-
-        Returns:
-            A list of validation issues and tasks needed for the graph.
-        """
-
-        mismatched_aliases = {}
-        swap_tasks: dict[str, Any] = {}
-        collected_no_tasks_found = []
-
-        for _step, task in pre_rendered_task_graph.items():
-            if isinstance(task, dict):
-                for swap_name, aliased_defns in task.items():
-                    if swap_name.startswith("?"):
-                        output_types = set()
-                        swap_dict = {}
-
-                        for alias, definition in aliased_defns.items():
-                            # figure out whether its long or short form
-                            if "task" in definition:
-                                # long version
-                                task_name = definition["task"]
-                            else:
-                                # short version - should be exactly one key in here
-                                task_name = list(definition.keys())[0]
-
-                            # get the output type of task_name in the plugin
-                            if task_name in task_lookup_dict:
-                                swap_dict[alias] = {
-                                    "plugin_snapshot_id": task_lookup_dict[task_name][
-                                        "plugin_snapshot_id"
-                                    ],
-                                    "pluginfile_filename": task_lookup_dict[task_name][
-                                        "pluginfile_filename"
-                                    ],
-                                    "task_name": task_lookup_dict[task_name][
-                                        "task_name"
-                                    ],
-                                }
-
-                                output_parameters: list[PluginTaskOutputParameter] = (
-                                    task_lookup_dict[task_name]["output_parameters"]
-                                )
-                                output_types.add(
-                                    tuple(
-                                        [
-                                            parameter.parameter_type.name
-                                            for parameter in output_parameters
-                                        ]
-                                    )
-                                )
-                            else:
-                                collected_no_tasks_found.append(
-                                    ValidationIssue(
-                                        type_=IssueType.SEMANTIC,
-                                        severity=IssueSeverity.ERROR,
-                                        message=f"In swap '{swap_name}', task with name '{task_name}' not found in registered tasks.",
-                                    )
-                                )
-
-                        swap_tasks[swap_name] = swap_dict
-
-                        if len(output_types) > 1:
-                            mismatched_aliases[swap_name] = output_types
-
-        return collected_no_tasks_found + [
-            ValidationIssue(
-                type_=IssueType.TYPE,
-                severity=IssueSeverity.ERROR,
-                message=f"Swap '{swap_name}' contains mismatched output types: {types}",
-            )
-            for swap_name, types in mismatched_aliases.items()
-        ], swap_tasks
-
-    def build_task_lookup_dict(
-        self, plugins: list[models.PluginPluginFile]
-    ) -> dict[str, Any]:
-        """Generates a lookup dictionary for function tasks.
-
-        Args:
-            plugins: A list of PluginPluginFile objects.
-
-        Returns:
-            A dictionary mapping function tasks to their plugin snapshot IDs, plugin filenames,
-            task names, and output parameters.
-        """
-        lookup = {}
-
-        for pair in plugins:
-            plugin = pair.plugin
-            file = pair.plugin_file
-
-            for task in file.tasks:
-                if isinstance(task, models.FunctionTask):
-                    # there must be a better way to check this
-
-                    lookup[task.plugin_task_name] = {
-                        "plugin_snapshot_id": plugin.resource_snapshot_id,
-                        "pluginfile_filename": file.filename,
-                        "task_name": task.plugin_task_name,
-                        "output_parameters": task.output_parameters,
-                    }
-
-        return lookup
-
-    def extract_swaps(self, task_graph: dict[str, Any]) -> dict[str, list[str]]:
-        """Extract all swaps from a task graph.
-
-        Args:
-            task_graph: The task graph dictionary.
-
-        Returns:
-            A dictionary mapping swap names to lists of available aliases.
-        """
-        swaps = {}
-
-        for _step, task in task_graph.items():
-            for swap_name, aliased_defns in task.items():
-                if swap_name.startswith("?"):
-                    swap_name_clean = swap_name[1:]  # Remove the '?' prefix
-                    if swap_name_clean not in swaps:
-                        swaps[swap_name_clean] = list(aliased_defns.keys())
-
-        return swaps
-
-    def validate_single_swap_combinations(
-        self,
-        task_graph_yaml: dict[str, Any],
-        artifact_graph: str,
-        plugin_plugin_files: list[models.PluginPluginFile],
-        plugin_parameter_types: list[models.PluginTaskParameterType],
-        entrypoint_parameters: list[dict[str, Any]],
-        entrypoint_artifacts: list[dict[str, Any]],
-        swap_choices: dict[str, Any],
-        log: BoundLogger,
-    ) -> tuple[list[ValidationIssue], set[str]]:
-        """Renders a graph containing swaps for a given set of swap_choices, and performs
-        validation on the rendered graph.
-
-        Args:
-            task_graph_yaml: The task graph dictionary.
-            artifact_graph: The artifact graph dictionary.
-            plugin_plugin_files: A list of Plugin-PluginFile objects.
-            plugin_parameter_types: A list of plugin parameter types.
-            entrypoint_parameters: A list of entrypoint parameters.
-            entrypoint_artifacts: A list of entrypoint artifacts parameters.
-            swap_choices: A dictionary representing the choices for each swap.
-            log: A BoundLogger instance.
-
-        Returns:
-            A list of validation issues and a list of global parameters required for the graph.
-        """
-        import json
-
-        from dioptra.restapi.v1.shared.entrypoint_validation import (
-            build_entrypoint_data_adapter,
-        )
-
-        issues_for_swap = []
-        required_globals: set[str] = set()
-
-        try:
-            rendered_graph = render_swaps_graph(task_graph_yaml, swap_choices)
-
-            entrypoint_data = build_entrypoint_data_adapter(
-                json.dumps(rendered_graph),
-                artifact_graph,
-                entrypoint_parameters,
-                entrypoint_artifacts,
-                log,
-            )
-            task_engine_dict = self._task_engine_yaml_service.build_dict(
-                entry_point=entrypoint_data,
-                plugin_plugin_files=plugin_plugin_files,
-                plugin_parameter_types=plugin_parameter_types,
-            )
-            # this is a schema check and deeper validation no longer present in workflows
-            issues = self._task_engine_yaml_service.validate(task_engine_dict)
-
-            for issue in issues:
-                issue.message = f"[Swap combination {swap_choices}] {issue.message}"
-                issues_for_swap.append(issue)
-
-            # collect any globals needed for this rendering
-            required_globals, _ = (
-                self._dynamic_global_parameter_service.get_required_globals(
-                    rendered_graph
-                )
-            )
-
-        except Exception as e:
-            issues_for_swap.append(
-                ValidationIssue(
-                    type_=IssueType.SEMANTIC,
-                    severity=IssueSeverity.ERROR,
-                    message=f"[Swap combination {swap_choices}] Error rendering graph: {str(e)}",
-                )
-            )
-
-        return issues_for_swap, required_globals
-
-
-class EntrypointSwapsValidationService(object):
-    """Initialize the swaps validation resource.
-
-    All arguments are provided via dependency injection.
-
-    Args:
-        swaps_validation_service: A SwapsValidationService object
-        entrypoint_id_service: An EntrypointIdService object.
-        entrypoint_snapshot_id_service: An EntrypointSnapshotIdService object.
-        task_engine_yaml_service: A TaskEngineYamlService object.
-    """
-
-    @inject
-    def __init__(
-        self,
-        swaps_validation_service: SwapsValidationService,
-        entrypoint_id_service: EntrypointIdService,
-        entrypoint_snapshot_id_service: EntrypointSnapshotIdService,
-        task_engine_yaml_service: TaskEngineYamlService,
-    ) -> None:
-        self._swaps_validation_service = swaps_validation_service
-        self._entrypoint_id_service = entrypoint_id_service
-        self._entrypoint_snapshot_id_service = entrypoint_snapshot_id_service
-        self._task_engine_yaml_service = task_engine_yaml_service
-
-    def validate(
-        self,
-        group_id: int,
-        swaps_graph: str,
-        artifact_graph: str,
-        entrypoint_parameters: list[dict[str, Any]],
-        entrypoint_artifacts: list[dict[str, Any]],
-        plugin_snapshot_ids: list[int],
-        **kwargs,
-    ) -> dict[str, Any]:
-        """Validation for a proposed entrypoint which may contain swaps, looping over swap combinations.
-
-        Args:
-            group_id: The group ID.
-            swaps_graph: The task graph dictionary.
-            artifact_graph: The artifact graph dictionary.
-            entrypoint_parameters: A list of entrypoint parameters.
-            entrypoint_artifacts: A list of entrypoint artifacts parameters.
-            plugin_snapshot_ids: A list of plugin snapshot IDs needed for the entrypoint.
-
-        Returns:
-            A ValidateSwapsResponseSchema object.
-        """
-
-        log: BoundLogger = kwargs.get("log", LOGGER.new())
-
-        swaps_yaml = yaml.safe_load(swaps_graph)
-
-        if swaps_yaml is None:
-            raise EmptyGraphError("Provided swaps graph is empty.")
-
-        # check for pre-render schema issues
-        schema_issues = self._swaps_validation_service.swaps_graph_validation(
-            pre_rendered_task_graph=swaps_yaml
-        )
-        schema_valid = len(schema_issues) == 0
-
-        
-        # Perform Swap Output Matching
-        plugin_plugin_files = get_plugin_plugin_files_from_plugin_snapshot_ids(
-            plugin_snapshot_ids=plugin_snapshot_ids, logger=log
-        )
-
-        # build a lookup dictionary for tasks from the plugin files
-        task_lookup_dict = self._swaps_validation_service.build_task_lookup_dict(
-            plugin_plugin_files
-        )
-
-        # validate that swap outputs match across the graph
-        output_issues, _ = self._swaps_validation_service.validate_swap_output_matches(
-            pre_rendered_task_graph=swaps_yaml,
-            task_lookup_dict=task_lookup_dict,
-        )
-
-        collected_rendered_validation_issues = []
-        collected_required_globals = set()
-
-        # no sense doing all this if the yaml was incorrect to begin with
-        if schema_valid:
-
-            swaps = self._swaps_validation_service.extract_swaps(
-                swaps_yaml
-            )  # { swap1: [alias1, alias2, alias3], swap2: [alias4, alias5, alias6], etc. }
-            swap_names = list(swaps.keys())  # [swap1, swap2, etc.]
-
-            combinations = []
-
-            for swap_name in swap_names:
-                # iterate over ONLY the current swap
-                for alias in swaps[swap_name]:
-                    current_swap_choice = {
-                        s: swaps[s][0] for s in swaps.keys() if s != swap_name
-                    }  # keeping the first swap from every *other* swap, as we vary the *current* swap, O(n) instead of O(w*x*y*z*...)!
-                    current_swap_choice[swap_name] = alias
-
-                    combinations.append(current_swap_choice)
-
-            combinations = [{}] if combinations == [] else combinations  #
-
-            for combination in combinations:
-                rendered_validation_issues, required_globals = (
-                    self._swaps_validation_service.validate_single_swap_combinations(
-                        artifact_graph=artifact_graph,
-                        task_graph_yaml=swaps_yaml,
-                        plugin_plugin_files=plugin_plugin_files,
-                        plugin_parameter_types=self._entrypoint_snapshot_id_service.get_group_plugin_parameter_types(
-                            group_id=group_id, logger=log
-                        ),
-                        entrypoint_parameters=entrypoint_parameters,
-                        entrypoint_artifacts=entrypoint_artifacts,
-                        swap_choices=combination,
-                        log=log,
-                    )
-                )
-                collected_required_globals.update(required_globals)
-                collected_rendered_validation_issues.extend(rendered_validation_issues)
-
-        # compile a set of all declared parameter names
-        declared_globals = {p["name"] for p in entrypoint_parameters}
-        declared_globals.update({a["name"] for a in entrypoint_artifacts})
-
-
-        missing_globals = [
-            g for g in collected_required_globals if g not in declared_globals
-        ]
-        return {
-            "schema_issues": schema_issues,
-            "swap_issues": output_issues,
-            "rendered_validation_errors": collected_rendered_validation_issues,
-            "missing_global_params": missing_globals,
-        }
